pyproject.toml
[project]
name = "genetique"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.85",
    "matplotlib>=3.10.7",
    "pandas>=2.2",
    "pytest>=8.0",
    "seaborn>=0.13.2",
]
